Record a response status on the AsyncResult object.
Some services that communicate via celery jobs may not send back a
result. In order to allow distinguishing successful responses from
failures, record the message status on the AyncResult object.

As part of this, make it clear whether a local timeout has occurred
by making this its own explicit state (this enables callers to do a
switch on the status and easily handle creation of an error in such
a case).