Add method to await worker job completion and use within tests.
Without waiting for the jobs to complete, it is possible that
disconnect in the after() hook might be called before a currently
executing job is finished with leads to an error once the job does
complete and storeResult() is called (because the backend redis
client will be disconnected).

This is required in order to allow testing timeout handling.