Harden the caching of previously awaited results.
There is no guarantee that upstream services will return a message
so the check to resolve with the result if it is already set is not
sufficient (the codepath was also missing return so the backend
would be called again).

Address both this and a race that existed while the backend request
was in flight but before it had resolved/rejected and assigned the
result by keeping a reference to the promise and returning that on
subsequent calls to the .get() method.