Record a response status on the AsyncResult object. (#25)
* Record a response status on the AsyncResult object.

Some services that communicate via celery jobs may not send back a
result. In order to allow distinguishing successful responses from
failures, record the message status on the AyncResult object.

As part of this, make it clear whether a local timeout has occurred
by making this its own explicit state (this enables callers to do a
switch on the status and easily handle creation of an error in such
a case).

* Convert timeouts and failure messages into rejections.

This commit removes the AsyncResult status property and instead
makes the calling convention resolution on success and rejection
on failure. The error message can be used to distinguish timeouts
from failure messages that are returned from the upstream system.

* Harden the caching of previously awaited results.

There is no guarantee that upstream services will return a message
so the check to resolve with the result if it is already set is not
sufficient (the codepath was also missing return so the backend
would be called again).

Address both this and a race that existed while the backend request
was in flight but before it had resolved/rejected and assigned the
result by keeping a reference to the promise and returning that on
subsequent calls to the .get() method.